split up legal move and tests into each piece.rb file and created test files
