+++ test/models/bishop_test.rb
@@ -1,0 +1,8 @@
+require 'test_helper'
+
+class BishopTest< ActiveSupport::TestCase
+	test "bishop_legal_move" do
+    	piece = FactoryGirl.create(:piece)
+    	assert piece.bishop_legal_move?(1, 1)
+	end
+end
