+++ test/models/game_test.rb
@@ -13,11 +13,5 @@ class GameTest < ActiveSupport::TestCase
   	actual = game.obstruction(0, 0).type
   	assert_equal expected, actual
   end
-
-  test "legal_move" do
-  	game = FactoryGirl.create(:game)
-  	expected = 'King'
-    actual = game.legal_move?(1, 1)
-  	assert_equal expected, actual
-  end
+  
 end
